Refactor coverage handling in Puppeteer integration tests
